move flow tester to flowlogs package
