+++ felix/fv/flow_logs_goldmane_test.go
@@ -37,7 +37,6 @@ import (
 	"github.com/projectcalico/calico/felix/fv/connectivity"
 	"github.com/projectcalico/calico/felix/fv/flowlogs"
 	"github.com/projectcalico/calico/felix/fv/infrastructure"
-	"github.com/projectcalico/calico/felix/fv/metrics"
 	"github.com/projectcalico/calico/felix/fv/utils"
 	"github.com/projectcalico/calico/felix/fv/workload"
 	"github.com/projectcalico/calico/libcalico-go/lib/apiconfig"
@@ -340,12 +339,12 @@ var _ = infrastructure.DatastoreDescribe("_BPF-SAFE_ goldmane flow log tests", [
 		// flow logs.
 		Eventually(func() error {
 			wepPort := 8055
-			flowTester := metrics.NewFlowTester(metrics.FlowTesterOptions{
+			flowTester := flowlogs.NewFlowTester(flowlogs.FlowTesterOptions{
 				ExpectLabels:           true,
 				ExpectEnforcedPolicies: true,
 				MatchEnforcedPolicies:  true,
 				MatchLabels:            false,
-				Includes:               []metrics.IncludeFilter{metrics.IncludeByDestPort(wepPort)},
+				Includes:               []flowlogs.IncludeFilter{flowlogs.IncludeByDestPort(wepPort)},
 				CheckNumFlowsStarted:   true,
 				CheckFlowsCompleted:    true,
 			})
@@ -355,7 +354,7 @@ var _ = infrastructure.DatastoreDescribe("_BPF-SAFE_ goldmane flow log tests", [
 				return fmt.Errorf("error populating flow logs from Felix[0]: %s", err)
 			}
 
-			aggrTuple := tuple.Make(flowlog.EmptyIP, flowlog.EmptyIP, 6, metrics.SourcePortIsNotIncluded, wepPort)
+			aggrTuple := tuple.Make(flowlog.EmptyIP, flowlog.EmptyIP, 6, flowlogs.SourcePortIsNotIncluded, wepPort)
 
 			host1_wl_Meta := endpoint.Metadata{
 				Type:           "wep",
