+++ felix/fv/tiered_policy_test.go
@@ -35,7 +35,6 @@ import (
 	"github.com/projectcalico/calico/felix/fv/connectivity"
 	"github.com/projectcalico/calico/felix/fv/flowlogs"
 	"github.com/projectcalico/calico/felix/fv/infrastructure"
-	"github.com/projectcalico/calico/felix/fv/metrics"
 	"github.com/projectcalico/calico/felix/fv/utils"
 	"github.com/projectcalico/calico/felix/fv/workload"
 	"github.com/projectcalico/calico/libcalico-go/lib/apiconfig"
@@ -459,7 +458,7 @@ var _ = infrastructure.DatastoreDescribe("connectivity tests and flow logs with 
 		})
 
 		checkFlowLogs := func() error {
-			aggrTuple := tuple.Make(flowlog.EmptyIP, flowlog.EmptyIP, 6, metrics.SourcePortIsNotIncluded, wepPort)
+			aggrTuple := tuple.Make(flowlog.EmptyIP, flowlog.EmptyIP, 6, flowlogs.SourcePortIsNotIncluded, wepPort)
 
 			host1_wl1_Meta := endpoint.Metadata{
 				Type:           "wep",
@@ -498,13 +497,13 @@ var _ = infrastructure.DatastoreDescribe("connectivity tests and flow logs with 
 				PortNum:   8066,
 			}
 
-			flowTester := metrics.NewFlowTester(metrics.FlowTesterOptions{
+			flowTester := flowlogs.NewFlowTester(flowlogs.FlowTesterOptions{
 				ExpectLabels:           true,
 				ExpectEnforcedPolicies: true,
 				MatchEnforcedPolicies:  true,
 				ExpectPendingPolicies:  true,
 				MatchPendingPolicies:   true,
-				Includes:               []metrics.IncludeFilter{metrics.IncludeByDestPort(wepPort)},
+				Includes:               []flowlogs.IncludeFilter{flowlogs.IncludeByDestPort(wepPort)},
 			})
 
 			err := flowTester.PopulateFromFlowLogs(tc.Felixes[0])
